pyproject.toml
[project]
name = "bootdevagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gitignore>=0.0.8",
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
